- null pointer fix - swagger config moved
